🔼 Refactor project structure and update .NET framework version
The changes include restructuring the 'Testing.Unit' and 'Testing.Unit.Facts' classes, updating them to use .NET 8.0, and changing the assembly information. The commit also includes the addition of new configuration files for the .idea folder. Unused or unneeded attributes, lines and files were also removed to streamline the projects.

+semver: breaking